requirements.txt
asgiref==3.6.0
Django==3.2.5
django-cors-headers==3.10.1
django-environ==0.10.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
